# docs/_archive/2026-09-18/openapi.yaml
openapi: 3.1.0
info:
  title: YYC³ NovaMind API
  version: 2.0.0
  description: |
    YYC³ NovaMind 全栈 API 契约（OpenAPI 3.1）。

    本契约由 M3 MiniMax 导师于 2026-09-18 自动扫描 `app/api/**` 导出函数生成，
    覆盖 13 个路由文件 / 30 个端点，作为前后端对接的唯一权威来源。

    ## 五维契约原则
    - **时间维**：所有时间字段统一 ISO-8601 UTC
    - **空间维**：所有路径相对 baseURL（默认 `http://localhost:3218`）
    - **属性维**：所有响应含 `success: boolean`，错误码与 `lib/error-handler.ts` 对齐
    - **事件维**：所有 POST/PUT 操作记录 trace_id 用于审计
    - **关联维**：复杂资源附带 `_links` 字段指向关联资源
  contact:
    name: YYC³ Team
    email: admin@0379.email
  license:
    name: Copyright © 2025-2026 YanYuCloudCube
    identifier: YYC-Internal
servers:
  - url: http://localhost:3218
    description: 本地开发（端口 3218）
  - url: https://staging.api.yyc3.app
    description: 预发布
  - url: https://api.yyc3.app
    description: 生产
...
paths:
...
  /api/speech-to-text:
    post:
      tags: [Speech]
      summary: 语音识别
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                language:
                  type: string
                  default: zh-CN
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags: [Speech]
      summary: 获取语音识别能力元数据
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ===== Status =====
...
  /api/sync:
    get:
      tags: [Sync]
      summary: 拉取增量数据
      parameters:
        - in: query
          name: since
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Sync]
      summary: 推送变更
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [Sync]
      summary: 合并远端数据
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Sync]
      summary: 删除同步条目
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ===== Offline Test =====
  /api/test-offline:
    get:
      tags: [Offline]
      summary: 离线模式探测
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Offline]
      summary: 模拟离线请求
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ===== Upload =====
...
components:
...
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          description: 业务是否成功
        data:
          description: 业务负载（任意 JSON）
        message:
          type: string
        trace_id:
          type: string
          description: 链路追踪 ID
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yyc3_session
      description: 会话 Cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 备用 Bearer Token
security:
  - sessionCookie: []
  - bearerAuth: []